+++ js/miniature
@@ -1,0 +1,21 @@
+const thumbnailTemplate = document.querySelector('#picture').content.querySelector('.picture');
+const picturesList = document.querySelector('.pictures');
+
+const createThumbnail = ({url, description, likes, comments}) => {
+  const thumbnail = thumbnailTemplate.cloneNode(true);
+  thumbnail.querySelector('.picture__img').src = url;
+  thumbnail.querySelector('.picture__img').alt = description;
+  thumbnail.querySelector('.picture__likes').textContent = likes;
+  thumbnail.querySelector('.picture__comments').textContent = comments.length;
+  return thumbnail;
+};
+
+const renderPictures = (photos) => {
+  const fragment = document.createDocumentFragment();
+  photos.forEach((photo) => {
+    fragment.appendChild(createThumbnail(photo));
+  });
+  picturesList.appendChild(fragment);
+};
+
+export { renderPictures };
